Parsing Offset Outputs and Data.

+++ HCP65PicoTester/Source/GeneralPins.h
@@ -7,8 +7,8 @@
 #define PIN_NOT_SET     (-1)
 
 
-uint32 make_8bit_mask(uint32* paiPins, uint32 iValue);
-uint32 make_4bit_mask(uint32* paiPins, uint32 iValue);
+uint32 make_8bit_mask(uint8* paiPins, uint8 iValue);
+uint32 make_4bit_mask(uint8* paiPins, uint8 iValue);
 
 
 #endif // __GENERAL_PINS_H__
